+++ lib/que/view.rb
@@ -12,7 +12,7 @@ module Que
     extend self
     extend Forwardable
 
-    # Public: Configure emailbutler.
+    # Public: Configure que view.
     #
     #   Que::View.configure do |config|
     #   end
@@ -26,15 +26,15 @@ module Que
       @configuration ||= Configuration.new
     end
 
-    # Public: Default per thread emailbutler instance if configured.
+    # Public: Default per thread que view instance if configured.
     # Returns Que::View::DSL instance.
     def instance
       Thread.current[:que_view_instance] ||= DSL.new
     end
 
     # Public: All the methods delegated to instance. These should match the interface of Que::View::DSL.
-    def_delegators :instance,
-                   :fetch_dashboard_stats, :fetch_failing_jobs, :fetch_scheduled_jobs, :fetch_job,
+    def_delegators :instance, :fetch_dashboard_stats,
+                   :fetch_running_jobs, :fetch_failing_jobs, :fetch_scheduled_jobs, :fetch_job,
                    :delete_failing_jobs, :delete_scheduled_jobs, :delete_job,
                    :reschedule_scheduled_jobs, :reschedule_failing_jobs, :reschedule_job
   end