+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- rendering running jobs
+
 ## [0.2.3] - 2023-12-01
 ### Fixed
-- jobs sttaus naming
+- jobs status naming
 
 ## [0.2.2] - 2023-11-15
 ### Added